+++ src/main/kotlin/com/fdtheroes/sgruntbot/handlers/message/oroscopo/HoroscopeParams.kt
@@ -1,0 +1,8 @@
+package com.fdtheroes.sgruntbot.handlers.message.oroscopo
+
+class HoroscopeParams(
+    val sign: Sign,
+    val planets: List<PlanetPosition>,
+    val conjunctions: List<List<PlanetPosition>>,
+    val oppositions: List<List<PlanetPosition>>,
+)
